added support for configuring a static multicast SA for VPLS mode
The static multicast SA introduces "multicast_group", "multicast_spi", and 
"multicast_key" options to hip.conf; peers must be configured with a matching
group/SPI/key combination.

When multicast packets are received from the overlay, instead of unicasting a
copy through each established tunnel, the static key/SPI is used and an ESP
packet is sent out to the configured group. This means multicast is sent
as multicast (multiple groups mapped to one), not as unicast, having different 
implications for how the ESP is routed over the underlay network.



git-svn-id: http://svn.code.sf.net/p/openhip/code/hip/trunk@260 4c5cb64f-9889-4596-9799-84b02dc3effa