Stop squatting on 1.0.0.0/8 for LSIs
It's a fully-valid chunk of the network, and what's more
it recently got reallocated for active use.

Instead of squatting on an active, general-use subnet, this
changes to squatting on 240.0.0.0/8 (the former "Class E"
addresses), which is "reserved for future addressing" as of
RFC 6890 (having obsoleted RFC 5735, which in turn obsoleted
RFC 3330, which copied the reservation from RFC 1700, which
was first added in RFC 990)

Since these have been reserved since 1986, they seem like
a safer bet than 1.0.0.0/8.